gameboard.js, .htaccess -- Finished new gamestate list retrieval
Finished the changes to gameboard.js (I had an issue with missing
parentheses on my xmlhttp send statement).  .htaccess was updated
to prevent caching of .json files; when the updateSaveButtons
function was called after a save, a cached version of the savegame
manifest would be retrieved even though it had since changed.