gameboard.php (v.), gameboard.js -- Fixed a save bug
Fixed a bug that caused saves of types "newgame" and "overwrite" to
fail.
KNOWN ISSUE: Select elements are not reset properly upon a save of
type "newgame"
WORKAROUND: Reload the page, and if that fails, navigate to
appliednerditry.com/chaos/gamelist.php (no need to do anything there,
just let it load) and then reload the board manager.