Begin implementing support for the Horned Rat CitOW expansion
The chaos.css stylesheet has been updated to reposition the free-
floating PP controls, to make way for the Horned Rat's figure box.
A new data file board_hr.xml has been created which includes
stats for The Horned Rat, and gameboard.js has been modified to
use this in place of board.xml for development (the production
code will allow for the use of either, depending on game type).
Also created (but untracked) is the 9797 savefile series, which
includes The Horned Rat, for use in development.