JavaScript updates to implement support for standard & expansion games
A new "blank" save game (saves/blankboard_hr.xml) was created for use
in starting new expansion games (not yet implemented).  This file
includes an 'expansion' attribute in the root tag, which will in the
future be propagated to additional expansion save files. The loader
functions were updated to automatically determine whether standard or
expansion game data is needed, and retrieve it accordingly.