Extended width of board and of regions, both dimensions of scoreboard
Each region's width was extended by 40px, to allow for two additional
columns of Old World tokens.  This allowed for an 80px extension of
the scoreboard, which was also expanded downward by 20px.  Dial
counter drawing was adjusted, and all upgrade icons were redrawn to
take advantage of the additional space.