Changed some tags in board data; minor edit to chaos cards
The tags "powers" and "player" in board.xml and board_hr.xml were
renamed to "ruinouspowers" and "power", respectively, to draw a
distinction between the similar, but non-identical "player" tags
in the save game files.

In chaos.xml and chaos_hr.xml, the placeholder card for The Horned One's
Due had the "owner" attribute added, and set to "Old_World".  This will
make it easier to identify (and keep) placeholder Chaos cards, while
screening out Chaos cards for players not in a game.