citow_draw.js -- Reduced board width, implemented figure shrinking
Region width was reduced by 20px (still 20px wider than the last
production release).  The scoreboard was reduced in width as well,
with most of that coming out of the name field.  Necessary
adjustments to gameboard.php and gameboard.js were also made.

In addition, a new feature was added: if a region becomes so crowded
with figures that the automatically calculated margin between groups
is negative, then the figure drawing area is shrunk (using the
transformations available on the HTML canvas 2D context).