newgame.php -- Call gamelist.php to update game manifests
An "include" statement invoking gamelist.php was inserted into
newgame.php just before the redirect to gameboard.php, to ensure that
the JSON game lists are updated.  A minor change to gamelist.php now
allows it to be called without echoing a response, to avoid issues
with "header('Location: ...')" redirects.