gameboard.php -- Version 8
The drawToken function's upgrade-drawing methods were reworked to
make use of the source image coordinates now stored in the upgrade
XML data.  The routine to load a board from local storage has been
modified so that the board will actually load only if the game
number matches the game the user is trying to access.  The upgrade
sprite sheets have been updated.