citow_draw.js -- Added third card slot (requires upgrade)
The region-drawing routine was updated to allow for a third Chaos
card to be drawn, if the card's owner possesses an upgrade with the
"extracard" property.  The game data was modified to give that
property to The Horned Rat's "thirteen" upgrade.  Gameboard.js was
updated to read the "extracard" upgrade property, and the drawBoard
routine was rearranged a bit so that regions are not drawn until
after player data (including upgrades) has been set up.

In addition, the game-saving routine was updated to (hopefully) save
player names; that has not yet been verified.