gameboard.js -- Prevent unwanted loading of a locally stored board
In gameboard.php, there is now a check to determine whether this
page was itself the referer; this information is stored in a hidden
input element, to be available to Javascript.  In gameboard.js, if
this same page was the referer, then a board in local storage will
not be loaded.

These changes are to ensure that the navigation controls at the
bottom of the page operate according to user expectations.  If the
user is deliberately navigating away from an unsaved board state,
they probably don't want to be forced right back to it.