Implemented a global reset button for figure effects
Using the same "button + arming toggle" idea as for the PP and DACs
reset buttons, an effects reset button was put into place.  The
PP reset was moved to the same box as the PP control buttons in
order to make room for it.  Changes to the CSS were required to
make everything look nice once again.