Removed over-board mousewheel PP adjustment
The ability to adjust PP via the mousewheel while hovering over the
actual drawn scoreboard was rendered obsolete long ago by the PP
controls box.  In addition, it caused inconsistent scrolling
behavior (since the page would scroll when the mouse was over any
other area of the board, but not the scoreboard).  Therefore, it
has been removed.