Fixed HTML/CSS workbox alignment bug in Opera
Opera renders the text for the page at a greater width than other
browsers do, which was pushing the toggle selectors and legends
out of the figure effects box ("workbox").  Once the box width
was corrected, it became apparent that Opera was not rendering the
text in the expected place, so a "vertical-align:top" declaration
was added as well.