Bug fix: un-break .png image export
While converting parts of the code to jQuery, image exporting was
inadvertenly broken due to the fact that jQuery properly sets a
MIME type on AJAX post data, and the original savepng.php code was
dependent on PHP's handling of data with no MIME type.  The script
savepng.php is updated to eliminate the issue.