Card interactions; fix missing GET vars bug
Refactor gameboard.php so that it doesn't error out when accessed
without a GET value for 'game'.

Implement the 'covercard' player upgrade feature, which allows a
Chaos card to replace (not displace) another Chaos card, when
dropped directly on top of it.  Handle interactions between this
feature and the existing 'extracard' feature.